Skip excluded files/dirs entirely.
Before this change: using -x to exclude a path would skip the content
of the file/dir, but still include its existance in the output tree.

After this change: file/dir is not added at all in the tree. This comes
at a cost, though, as it can't match on the path anymore. Only the file
name.

Additionally, exclude .swp files by default, and updated test output.