Address another instance of tag not being identified.
When a construct was identified as a declaration, we only turned it into
a linkable element if first declaration and declaration matched id.
Not sure why, probably a workaround for an old forgotten bug.

This caused a few data structures not to be wrapped with the correct id.
In this change, fixed it.