Bugfix: kaslr.c was causing sbexr to crash.
Background: one of the includes in kaslr.c had a #if .. #include .. #else
pattern. For some reason, it confused the code to determine if a header
had to be parsed or not, causing the part of the file before the include
to be skipped, and the part after the include to be parsed, and a crash
at the #else, which found an empty #if stack.

In this change:
- deprecate the old logic in pp-tracker.h using a stack and an
  heuristic of dubious correctness with a much simpler, more
  understandable heuristic of similarly dubious correcctness which
  however seem to be working correctly on all the input cases.